+++ internal/policy/model.go
@@ -136,10 +136,11 @@ type Project struct {
 }
 
 type Scope struct {
-	PythonContracts []PythonContract `json:"pythonContracts,omitempty"`
-	Exclude         []string         `json:"exclude,omitempty"`
-	Generated       []string         `json:"generated,omitempty"`
-	Data            []string         `json:"data,omitempty"`
+	PythonRuntimeLoaders []PythonRuntimeLoader `json:"pythonRuntimeLoaders,omitempty"`
+	PythonContracts      []PythonContract      `json:"pythonContracts,omitempty"`
+	Exclude              []string              `json:"exclude,omitempty"`
+	Generated            []string              `json:"generated,omitempty"`
+	Data                 []string              `json:"data,omitempty"`
 
 	EntryPoints                 []string                     `json:"entryPoints,omitempty"`
 	GeneratedJavaScript         []GeneratedJavaScript        `json:"generatedJavaScript,omitempty"`